% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_distance_matrix}
\alias{calculate_distance_matrix}
\title{Calculate Distance Matrix Between Locations}
\usage{
calculate_distance_matrix(spatial_obj, units = "km")
}
\arguments{
\item{spatial_obj}{sf object with spatial geometries}

\item{units}{Character. "km" or "m". Default "km".}
}
\value{
Distance matrix
}
\description{
Computes pairwise distances between facilities or districts.
}
