% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_org_group}
\alias{classify_org_group}
\title{Classify Organism Group}
\usage{
classify_org_group(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame}

\item{organism_col}{Character. Normalized organism column.
Default "organism_normalized".}
}
\value{
Data frame with org_group column added
}
\description{
Assigns organisms to taxonomic groups (Enterobacterales, Gram-positive, etc.)
}
