% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{convert_df_to_html_table}
\alias{convert_df_to_html_table}
\title{Convert Data Frame to HTML Table}
\usage{
convert_df_to_html_table(df)
}
\description{
Convert Data Frame to HTML Table
}
\keyword{internal}
