% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daly_resistance_prevalence.R
\name{daly_map_rr_drug_class}
\alias{daly_map_rr_drug_class}
\title{Map Antibiotic Class to RR Drug Category}
\usage{
daly_map_rr_drug_class(data, class_col = "antibiotic_class")
}
\arguments{
\item{data}{Data frame}

\item{class_col}{Character. Antibiotic class column. Default "antibiotic_class".}
}
\value{
Data frame with rr_drug column added
}
\description{
Maps WHO antibiotic classes to RR drug categories for burden estimation.
}
