% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden.R
\name{.compute_mean_from_fit}
\alias{.compute_mean_from_fit}
\title{Compute analytical mean LOS from a fitdistrplus fit object}
\usage{
.compute_mean_from_fit(fit, dist)
}
\description{
Compute analytical mean LOS from a fitdistrplus fit object
}
\keyword{internal}
