% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_report_html}
\alias{export_report_html}
\title{Export Report as HTML}
\usage{
export_report_html(report, file)
}
\arguments{
\item{report}{Report object}

\item{file}{Output file path}
}
\description{
Export Report as HTML
}
\keyword{internal}
