% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_report_pdf}
\alias{export_report_pdf}
\title{Export Report as PDF}
\usage{
export_report_pdf(report, file)
}
\arguments{
\item{report}{Report object}

\item{file}{Output file path}
}
\description{
Export Report as PDF
}
\keyword{internal}
