% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_report_txt}
\alias{export_report_txt}
\title{Export Report as Text}
\usage{
export_report_txt(report, file)
}
\arguments{
\item{report}{Report object}

\item{file}{Output file path}
}
\description{
Export Report as Text
}
\keyword{internal}
