% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{extract_species}
\alias{extract_species}
\title{Extract Species from Organism Name}
\usage{
extract_species(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame}

\item{organism_col}{Character. Organism column name.
Default "organism_normalized".}
}
\value{
Data frame with org_species column added
}
\description{
Extracts bacterial species (second word) from normalized organism name.
}
