% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_amr_palette}
\alias{get_amr_palette}
\title{Get Color Palette}
\usage{
get_amr_palette(palette_name = "default")
}
\arguments{
\item{palette_name}{Character. Name of palette: "resistance", "outcome", "default", "green", "spectral".}
}
\value{
Named vector of colors
}
\description{
Returns predefined color palettes for AMR visualizations.
}
