% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clean_and_standardize.R
\name{get_antimicrobial_categories}
\alias{get_antimicrobial_categories}
\title{Get Antimicrobial Categories for MDR/XDR Classification}
\usage{
get_antimicrobial_categories(organism_group = "Enterobacterales")
}
\arguments{
\item{organism_group}{Character. Organism group name. Default
\code{"Enterobacterales"}.}
}
\value{
Character vector of antimicrobial category names.
}
\description{
Returns the antimicrobial categories used for Magiorakos MDR/XDR
classification for a given organism group.
}
