% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_beta_lactam_hierarchy}
\alias{get_beta_lactam_hierarchy}
\title{Beta-Lactam Class Hierarchy}
\usage{
get_beta_lactam_hierarchy()
}
\value{
Character vector
}
\description{
Ordered vector of beta-lactam classes for resistance class selection.
Order represents clinical hierarchy (most to least important).
}
