% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{map_class_to_rr}
\alias{map_class_to_rr}
\title{Map Antibiotic Class to RR Drug Category}
\usage{
map_class_to_rr(data, class_col = "antibiotic_class")
}
\arguments{
\item{data}{Data frame}

\item{class_col}{Character. Antibiotic class column. Default "antibiotic_class".}
}
\value{
Data frame with rr_drug column added
}
\description{
Maps WHO antibiotic classes to RR drug categories for burden estimation.
}
