% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clean_and_standardize.R
\name{prep_assign_organism_group}
\alias{prep_assign_organism_group}
\title{Assign Organism Group}
\usage{
prep_assign_organism_group(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame}

\item{organism_col}{Character. Normalized organism column.
Default "organism_normalized".}
}
\value{
Data frame with org_group column added
}
\description{
Assigns organisms to taxonomic groups (Enterobacterales, Gram-positive, etc.)
}
