% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clean_and_standardize.R
\name{prep_calculate_los}
\alias{prep_calculate_los}
\title{Calculate Length of Stay}
\usage{
prep_calculate_los(
  data,
  admission_col = "date_of_admission",
  outcome_col = "date_of_final_outcome"
)
}
\arguments{
\item{data}{Data frame}

\item{admission_col}{Character. Admission date column.}

\item{outcome_col}{Character. Outcome/discharge date column.}
}
\value{
Data frame with Length_of_stay column
}
\description{
Calculates hospital length of stay in days.
}
