% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hai_cai.R
\name{prep_derive_icu_flag}
\alias{prep_derive_icu_flag}
\title{Derive ICU Flag}
\usage{
prep_derive_icu_flag(
  data,
  ward_col = "ward_icu",
  dept_col = "hospital_department"
)
}
\arguments{
\item{data}{Data frame.}

\item{ward_col}{Character. Ward/unit column. Default "ward_icu".}

\item{dept_col}{Character. Department column. Default "hospital_department".}
}
\value{
Data frame with \code{icu_flag} logical column added.
}
\description{
Sets \code{icu_flag = TRUE} where ward/department information indicates
the patient was in an ICU at the time of culture.
}
