% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clean_and_standardize.R
\name{prep_extract_genus}
\alias{prep_extract_genus}
\title{Extract Genus from Organism Name}
\usage{
prep_extract_genus(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame}

\item{organism_col}{Character. Organism column name.
Default "organism_normalized".}
}
\value{
Data frame with org_genus column added
}
\description{
Extracts bacterial genus (first word) from normalized organism name.
}
