% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ast_and_syndrome.R
\name{prep_standardize_ast_values}
\alias{prep_standardize_ast_values}
\title{Standardize Susceptibility Values}
\usage{
prep_standardize_ast_values(data, col = "antibiotic_value")
}
\arguments{
\item{data}{Data frame containing susceptibility column}

\item{col}{Character. Name of susceptibility column. Default "antibiotic_value".}
}
\value{
Data frame with standardized susceptibility values and added column
  \code{antibiotic_value_std}
}
\description{
Maps various susceptibility/resistance representations to standard
"S" (Susceptible), "R" (Resistant), "I" (Intermediate) values.
}
