% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_optional_columns.R
\name{prep_standardize_sex}
\alias{prep_standardize_sex}
\title{Standardize Sex Values}
\usage{
prep_standardize_sex(data, col = "gender")
}
\arguments{
\item{data}{Data frame containing gender column.}

\item{col}{Character. Name of gender column. Default "gender".}
}
\value{
Data frame with standardized gender values.
}
\description{
Maps various gender/sex representations to standard "M" or "F".
}
