% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.amr_config}
\alias{print.amr_config}
\title{Print AMR Configuration}
\usage{
\method{print}{amr_config}(x, ...)
}
\arguments{
\item{x}{An amr_config object}

\item{...}{Additional arguments (not used)}
}
\description{
Print AMR Configuration
}
