% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.amr_preprocessing_report}
\alias{print.amr_preprocessing_report}
\title{Print Preprocessing Report}
\usage{
\method{print}{amr_preprocessing_report}(x, ...)
}
\arguments{
\item{x}{An amr_preprocessing_report object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Preprocessing Report
}
