% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shorten_drug_class}
\alias{shorten_drug_class}
\title{Shorten Antibiotic Class Names}
\usage{
shorten_drug_class(x)
}
\arguments{
\item{x}{Character vector of antibiotic class names.}
}
\value{
Character vector of shortened names.
}
\description{
Maps long antibiotic class names to common abbreviations used in GBD-style
figures.
}
\examples{
shorten_drug_class(c("Carbapenems", "Third-generation-cephalosporins"))
}
