% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{standardize_gender}
\alias{standardize_gender}
\title{Standardize Gender Values}
\usage{
standardize_gender(data, col = "gender")
}
\arguments{
\item{data}{Data frame containing gender column}

\item{col}{Character. Name of gender column. Default "gender".}
}
\value{
Data frame with standardized gender values
}
\description{
Maps various gender representations to standard "M" or "F" values.
}
