% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{standardize_outcome}
\alias{standardize_outcome}
\title{Standardize Outcome Values}
\usage{
standardize_outcome(data, col = "final_outcome")
}
\arguments{
\item{data}{Data frame containing outcome column}

\item{col}{Character. Name of outcome column. Default "final_outcome".}
}
\value{
Data frame with standardized outcome values
}
\description{
Maps various outcome representations to standard values:
"Died", "Discharged", "LAMA", "Unknown"
}
