% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summarize_raw_data}
\alias{summarize_raw_data}
\title{Summarize Raw Data}
\usage{
summarize_raw_data(data)
}
\arguments{
\item{data}{Raw input data frame}
}
\value{
List with raw data summaries
}
\description{
Creates summary statistics for the original input dataset.
}
