% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summarize_transformations}
\alias{summarize_transformations}
\title{Summarize Data Transformations}
\usage{
summarize_transformations(raw_data, processed_data, preprocessing_log)
}
\arguments{
\item{raw_data}{Original data}

\item{processed_data}{Processed data}

\item{preprocessing_log}{Full preprocessing log}
}
\value{
List with transformation summaries
}
\description{
Documents all transformations applied to data values.
}
