% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summarize_value_mapping}
\alias{summarize_value_mapping}
\title{Summarize Value Mapping}
\usage{
summarize_value_mapping(raw_data, processed_data, raw_col, processed_col)
}
\description{
Summarize Value Mapping
}
\keyword{internal}
