% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_column_mappings}
\alias{default_column_mappings}
\title{Default Column Name Mappings}
\format{
A named list with 15 elements
}
\usage{
default_column_mappings
}
\description{
Named list mapping standard column names to common aliases found in
AMR surveillance datasets from different sources.
}
\keyword{datasets}
