% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_schema.R
\docType{data}
\name{default_column_mappings}
\alias{default_column_mappings}
\title{Default column name mappings for fuzzy matching}
\format{
A named list with elements mapping standard names to character
  vectors of acceptable aliases.
}
\usage{
default_column_mappings
}
\description{
Default column name mappings for fuzzy matching
}
\keyword{datasets}
