% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{.check_cols}
\alias{.check_cols}
\title{Validate that required columns exist in a data frame}
\usage{
.check_cols(data, cols)
}
\description{
Validate that required columns exist in a data frame
}
\keyword{internal}
