% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_antimicrobial_categories}
\alias{get_antimicrobial_categories}
\title{Get Antimicrobial Categories for MDR/XDR Classification}
\usage{
get_antimicrobial_categories(organism_group = "Enterobacterales")
}
\arguments{
\item{organism_group}{Character. Organism group name.}
}
\value{
Character vector of antimicrobial categories for that organism
}
\description{
Returns antimicrobial categories used in Magiorakos MDR/XDR definitions.
Categories are pathogen-specific.
}
