% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_standardize_organisms.R
\name{prep_extract_genus}
\alias{prep_extract_genus}
\title{Extract Genus from Organism Name}
\usage{
prep_extract_genus(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame.}

\item{organism_col}{Character. Organism column. Default "organism_normalized".}
}
\value{
Data frame with org_genus column added.
}
\description{
Extract Genus from Organism Name
}
