% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_standardize_organisms.R
\name{prep_extract_species}
\alias{prep_extract_species}
\title{Extract Species from Organism Name}
\usage{
prep_extract_species(data, organism_col = "organism_normalized")
}
\arguments{
\item{data}{Data frame.}

\item{organism_col}{Character. Organism column. Default "organism_normalized".}
}
\value{
Data frame with org_species column added.
}
\description{
Extract Species from Organism Name
}
